fix: merge AppRoutes in App component + minor yarn.lock update :art:
